Moving out dependencies of CusomEditor and the component itself outside, and refactor
